Hotfix: dse command no longer deletes all se's

+++ src/main/java/net/purplegoose/didnb/commands/scheduled_events/DeleteAllEvents.java
@@ -3,7 +3,9 @@ package net.purplegoose.didnb.commands.scheduled_events;
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.ScheduledEvent;
+import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.purplegoose.didnb.DiabloImmortalDiscordNotifier;
 import net.purplegoose.didnb.annotations.CommandAnnotation;
 import net.purplegoose.didnb.commands.Command;
 import net.purplegoose.didnb.data.LoggingInformation;
@@ -26,7 +28,10 @@ public class DeleteAllEvents extends Command {
             return;
         }
         for (ScheduledEvent scheduledEvent : guild.getScheduledEvents()) {
-            scheduledEvent.delete().queue();
+            User creator = scheduledEvent.getCreator();
+            if (creator != null && creator.getId().equalsIgnoreCase(DiabloImmortalDiscordNotifier.botId)) {
+                scheduledEvent.delete().queue();
+            }
         }
         createUseLogEntry(logInfo, DeleteAllEvents.class);
         event.getHook().setEphemeral(true).sendMessage("Deleting all scheduled messages...").queue();
